requirements.txt
python-telegram-bot[job-queue]==20.3
aiohttp
matplotlib
pandas
openpyxl
